Ajustes en flujos de eliminación

+++ src/main/java/application/core/services/CursoService.java
@@ -4,14 +4,19 @@ import java.util.List;
 
 import application.core.interfaces.ICurso;
 import global.Result;
+import infrastructure.core.interfaces.IAlumnoRepository;
 import infrastructure.core.interfaces.ICursoRepository;
+import infrastructure.core.interfaces.IMatriculaRepository;
+import infrastructure.core.models.Alumno;
 import infrastructure.core.models.Curso;
 
 public class CursoService implements ICurso {
     private final ICursoRepository _cursoRepository;
+    private final IAlumnoRepository _alumnoRepository;
 
-    public CursoService(ICursoRepository cursoRepository) {
+    public CursoService(ICursoRepository cursoRepository, IAlumnoRepository alumnoRepository) {
         _cursoRepository = cursoRepository;
+        _alumnoRepository = alumnoRepository;
     }
 
     public Result<List<Curso>> getAllCursos() {
@@ -31,6 +36,16 @@ public class CursoService implements ICurso {
     }
 
     public Result<Void> deleteCurso(int codigo) {
+        Result<List<Alumno>> alumnosMatriculados = _alumnoRepository.getAlumnosMatriculadosEnCurso(codigo);
+
+        if (alumnosMatriculados.isError()) {
+            return Result.error(alumnosMatriculados.getError());
+        }
+
+        if (!alumnosMatriculados.getValue().isEmpty()) {
+            return Result.error("No se puede eliminar el curso ya que hay alumnos matriculados en él");
+        }
+
         return _cursoRepository.deleteCurso(codigo);
     }
